requirements.txt
fastapi
uvicorn
jinja2
python-multipart
sentence-transformers
faiss-cpu
pandas
python-docx
passlib[bcrypt]
python-jose
sqlalchemy
